whisper-asr-server/requirements.txt
# Whisper ASR Server Requirements
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
faster-whisper>=0.9.0
whisper-timestamped>=1.14.0
soundfile>=0.12.1
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.24.0
aiofiles>=23.0.0
# Audio processing
pydub>=0.25.1
# Model caching
huggingface-hub>=0.16.0
# Monitoring
psutil>=5.9.0